--- app/src/main/java/ru/practicum/android/diploma/search/data/mapper/SearchResultMapper.kt
@@ -1,11 +1,0 @@
-package ru.practicum.android.diploma.search.data.mapper
-
-import ru.practicum.android.diploma.search.data.model.HHApiResponse
-import ru.practicum.android.diploma.search.domain.model.SearchResult
-
-fun HHApiResponse.Vacancies.toSearchResult() = SearchResult(
-    resultsFound = this.found,
-    totalPages = this.pages,
-    currentPage = this.page,
-    vacancies = this.items.map { it.toVacancy() },
-)
